+++ libc/src/__support/CMakeLists.txt
@@ -493,6 +493,7 @@ add_subdirectory(net)
 add_subdirectory(threads)
 
 add_subdirectory(File)
+add_subdirectory(pwd)
 
 add_subdirectory(printf_core)
