+++ libc/test/src/__support/pwd/flat_file_db_test.cpp
@@ -14,8 +14,8 @@
 #include "hdr/errno_macros.h"
 #include "src/__support/CPP/span.h"
 #include "src/__support/File/file.h"
-#include "src/pwd/field_tokenizer.h"
-#include "src/pwd/flat_file_db.h"
+#include "src/__support/pwd/field_tokenizer.h"
+#include "src/__support/pwd/flat_file_db.h"
 #include "src/stdio/remove.h"
 #include "src/string/string_utils.h"
 #include "test/UnitTest/ErrnoCheckingTest.h"
