+++ libc/test/src/pwd/CMakeLists.txt
@@ -5,33 +5,6 @@ if(NOT TARGET libc.src.pwd.pwd_utils)
 endif()
 
 add_libc_test(
-  field_tokenizer_test
-  SUITE
-    libc_pwd_unittests
-  SRCS
-    field_tokenizer_test.cpp
-  DEPENDS
-    libc.src.__support.CPP.span
-    libc.src.pwd.pwd_utils
-)
-
-add_libc_test(
-  flat_file_db_test
-  SUITE
-    libc_pwd_unittests
-  SRCS
-    flat_file_db_test.cpp
-  DEPENDS
-    libc.hdr.errno_macros
-    libc.src.__support.CPP.span
-    libc.src.__support.File.file
-    libc.src.__support.File.platform_file
-    libc.src.pwd.pwd_utils
-    libc.src.stdio.remove
-    libc.src.string.string_utils
-)
-
-add_libc_test(
   pwd_utils_test
   SUITE
     libc_pwd_unittests
